converted int amounts to decimal

+++ RDS.ExpenseTracker.Api/Dtos/FinancialAccountDto.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string? Name { get; set; }
-        public int? Availability { get; set; }
+        public decimal? Availability { get; set; }
         public string? Description { get; set; }
     }
 }
